timers: Split posix timers restore into two stages
1st stage is -- creating the timers. It may fail if kernel
allocated IDs in a manner we don't expect or runs out of
memory.

2nd stage is -- arm the timers. It cannot fail, since we've
validated the timespecs in advance and should happen after
we've waited for all the other tasks to complete the restore.

+++ include/restorer.h
@@ -64,7 +64,7 @@ struct rst_sched_param {
 };
 
 struct str_posix_timer {
-	int it_id;
+	long it_id;
 	int clock_id;
 	int si_signo;
 	int it_sigev_notify;
